added parsing fields support

+++ src/logsort/logsort/Program.cs
@@ -2,11 +2,8 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Net.Mime;
-using System.Reflection;
 using System.Text;
-using System.Threading.Tasks;
+using System.Text.RegularExpressions;
 using CliParse;
 
 namespace logsort
@@ -19,9 +16,48 @@ namespace logsort
             [CliParse.ParsableArgument("path", ShortName = 'p', ImpliedPosition = 0, Required = true)]
             public string Path { get; set; }
 
-            [CliParse.ParsableArgument("out", ShortName = 'o', DefaultValue = "logsort_output.txt")]
+            [CliParse.ParsableArgument("match", ShortName = 'm', DefaultValue = ".*")]
+            public string InputFileMatch { get; set; }
+
+            [CliParse.ParsableArgument("keymatch", ShortName = 'k', DefaultValue = "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2},\\d{3}")]
+            public string KeyMatch { get; set; }
+
+            [CliParse.ParsableArgument("prefix", ShortName = 'f', DefaultValue = "")]
+            public string EntryPrefix { get; set; }
+
+            [CliParse.ParsableArgument("entrymatch", ShortName = 'e', DefaultValue = "")]
+            public string EntryMatch { get; set; }
+
+            [CliParse.ParsableArgument("entryinclude", ShortName = 'i', DefaultValue = "")]
+            public string EntryIncludes { get; set; }
+            
+            [CliParse.ParsableArgument("out", ShortName = 'o', DefaultValue = "")]
             public string OutputFile { get; set; }
-             
+
+            public bool ShouldAddEntry(string entry)
+            {
+                if (EntryMatch == "" && EntryIncludes == "") return true;
+
+                if (EntryIncludes != "")
+                {
+                    if (entry.IndexOf(EntryIncludes, StringComparison.InvariantCultureIgnoreCase) > -1) return true;
+                }
+
+                if (EntryMatch == "") return false;  // only attempt to regex match if we have a pattern
+                var isMatch = Regex.IsMatch(entry, EntryMatch);
+                return isMatch;
+
+            }
+
+            public string GetKeyFromLine(string line)
+            {
+                return Regex.Match(line, KeyMatch).Value;
+            }
+
+            public bool IsStartOfEntry(string line)
+            {
+                return Regex.IsMatch(line, KeyMatch);
+            }
         }
 
         public class DuplicateKeyComparer<TKey>:IComparer<TKey> where TKey : IComparable
@@ -29,9 +65,7 @@ namespace logsort
             public int Compare(TKey x, TKey y)
             {
                 int result = x.CompareTo(y);
-
                 if (result == 0) return 1;
-                
                 return result;
             }
         }
@@ -52,59 +86,74 @@ namespace logsort
                     return;
                 }
 
-                sortConfig.OutputFile = Path.Combine(sortConfig.Path, sortConfig.OutputFile);
-                
                 var logentries = new SortedList<string, string>(new DuplicateKeyComparer<string>());
 
                 string key = "";
+                var entry = "";
                 var sb = new StringBuilder();
-                foreach (var file in Directory.EnumerateFiles(sortConfig.Path))
+
+                // SCAN
+
+                // figure out if we are sorting a single file or a folder of files.
+                IEnumerable<string> enumerateFiles;
+                if (File.Exists(sortConfig.Path))
+                {
+                    enumerateFiles = new String[] {sortConfig.Path};
+                }
+                else
+                {
+                    enumerateFiles=Directory.EnumerateFiles(sortConfig.Path);
+                }
+                foreach (var file in enumerateFiles)
                 {
                     if(file.Equals(sortConfig.OutputFile,StringComparison.InvariantCultureIgnoreCase)) continue; // skip our output file in case it already exists.
+                    if(!Regex.IsMatch(file,sortConfig.InputFileMatch)) continue;  // skip files we are ignoring
 
                     foreach (var line in File.ReadLines(file))
                     {
-                        if (IsStartOfEntry(line))
+                        if (sortConfig.IsStartOfEntry(line))
                         {
-                            // add entry
-                            logentries.Add(key, sb.ToString());
+                            if (key == "") key = sortConfig.GetKeyFromLine(line);
+
+                            entry = sb.ToString();
+                            // check if we should add entry.
+                            if (sortConfig.ShouldAddEntry(entry)) logentries.Add(key, entry);
                             
                             sb = new StringBuilder();
                             // set new key
-                            key = GetKeyFromLine(line);
-                            sb.AppendLine(line.Trim());
+                            key = sortConfig.GetKeyFromLine(line);
+                            sb.Append(line.Trim());
                             continue;
                         }
-                        sb.AppendLine("    " + line.Trim());
-                    }
-
-                    logentries.Add(key, sb.ToString());
+                        sb.Append("\n" + sortConfig.EntryPrefix + line.TrimEnd( '\r', '\n' ));
+                   }
+                    
+                    // check if we should add entry.
+                    entry = sb.ToString();
+                    if(entry.Length>0 && sortConfig.ShouldAddEntry(entry)) logentries.Add(key, entry);
                     sb = new StringBuilder();
                 }
-                
-                File.Delete(sortConfig.OutputFile);
-                File.WriteAllLines(sortConfig.OutputFile, logentries.Values);
+
+
+                // OUTPUT
+                if (sortConfig.OutputFile != "")
+                {
+                    File.Delete(sortConfig.OutputFile);
+                    File.WriteAllLines(sortConfig.OutputFile, logentries.Values);
+                }
+                else
+                {
+                    foreach (var logentry in logentries.Values)
+                    {
+                        Console.WriteLine(logentry);
+                    }
+                }
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
                 Console.WriteLine(ex.StackTrace);
             }
-        }
-
-        private static string GetKeyFromLine(string line)
-        {
-            if (line.Length < 23) return "";
-
-            return line.Substring(0, 23);
-        }
-
-        private static bool IsStartOfEntry(string line)
-        {
-            if (string.IsNullOrEmpty(line)) return false;
-            if (string.IsNullOrEmpty(line.TrimStart())) return false;
-
-            return line.TrimStart()[0] == '2';// entries start with a timestamp '2015....'
         }
     }
 }
